requirements.txt
flask>=2.0.0
psycopg2-binary>=2.9.0
faker>=18.0.0